Use Bootstrap collapsible navbar for smaller screens
Gives us a proper way to show breadcrumbs on smaller screens
instead of us having to manually create a dropdown out of the
breadcrumb elements.

Besides, dropdown items have to have the dropdown-item class in
Bootstrap 5 to have proper styling. Since we're rendering the
breadcrumbs exactly the same whether they're going to appear in a
breadcrumb or not, it's not simple (at least with the current
implementation) to dynamically add or remove the dropdown-item
class. Therefore, an alternative had to be found.