Make right-hand side elements in the navbar always visible
Although breadcrumbs appear last in the code, order-lg-# classes
are used to specify they should go first on large screens. This is
to avoid Bootstrap navbar hiding the avatar, level and notification
bell icons when collapsing, which it does if they're declared after
the collapsible components (the breadcrumbs in this case).