Give proper styles to checkboxes
Add form-check to parent divs, form-check-input for checkboxes
and form-check-label (if needed) to labels. For read-only editor,
add mu-form (even though it's not a form) for styling. Also, I'm
keeping the default checkbox color (primary/blue) instead of
forcing complementary (pink).

I sense it'll be an unpopular decision I'll have to revert :-)