Consider mu-kids-context might not be present when initializing
A modal for mu-kids-context is shown when mumuki.Kids is created,
as .showContext() is called on the constructor. While a
mu-kids-context *is* expected on real-case scenarios, I'm adding a
fail-safe as certain tests not related to the context at all were
failing because the modal couldn't be found when initializing
mumuki.Kids.