Add class for each filtered discussion's <a> tag on index discussions
Most organizations change the link colors through custom CSS.
Since each discussion on the index is wrapped in an <a> tag,
all text would change to the link color, which doesn't look very
good. Now that the tag can be easily identified, link colors can
be changed using a:not(.discussion-link).