Remove organization from search
It has no sense. An exam authorization always belongs to a
registration, which in turns belong to an organization. Although it may
be a good idea to denormalize that field for quering reasons, by no ways
we should assume it will matter when looking for an authorization to a
already known registration